+++ udata/core/badges/models.py
@@ -1,14 +1,15 @@
 import logging
-import weakref
 
 from datetime import datetime
 
 from mongoengine.signals import post_save
 
+from udata.api_fields import field
 from udata.auth import current_user
 from udata.mongo import db
 
 from .signals import on_badge_added, on_badge_removed
+from .fields import badge_fields
 
 log = logging.getLogger(__name__)