+++ udata/core/spatial/api.py
@@ -85,7 +85,7 @@ dataset_parser.add_argument(
     location='args', default=25)
 
 
-@ns.route('/zones/<pathlist:ids>/', endpoint='zones')
+@ns.route('/zones/<list:ids>/', endpoint='zones')
 class ZonesAPI(API):
     @api.doc('spatial_zones',
              params={'ids': 'A zone identifiers list (comma separated)'})
@@ -101,7 +101,7 @@ class ZonesAPI(API):
         }
 
 
-@ns.route('/zone/<path:id>/datasets/', endpoint='zone_datasets')
+@ns.route('/zone/<id>/datasets/', endpoint='zone_datasets')
 class ZoneDatasetsAPI(API):
     @api.doc('spatial_zone_datasets', params={'id': 'A zone identifier'})
     @api.expect(dataset_parser)
@@ -118,7 +118,7 @@ class ZoneDatasetsAPI(API):
         return datasets
 
 
-@ns.route('/zone/<path:id>/', endpoint='zone')
+@ns.route('/zone/<id>/', endpoint='zone')
 class ZoneAPI(API):
     @api.doc('spatial_zone', params={'id': 'A zone identifier'})
     def get(self, id):
@@ -152,7 +152,7 @@ class SpatialGranularitiesAPI(API):
         } for id, name in spatial_granularities]
 
 
-@ns.route('/coverage/<path:level>/', endpoint='spatial_coverage')
+@ns.route('/coverage/<level>/', endpoint='spatial_coverage')
 class SpatialCoverageAPI(API):
     @api.doc('spatial_coverage')
     @api.marshal_list_with(feature_collection_fields)