+++ udata/tests/api/test_organizations_api.py
@@ -191,8 +191,7 @@ class MembershipAPITest:
         user = api.login()
         data = {'comment': 'a comment'}
 
-        api_url = url_for('api.request_membership', org=organization)
-        response = api.post(api_url, data)
+        response = api.post(url_for('api.request_membership', org=organization), data)
         assert201(response)
 
         organization.reload()
@@ -209,14 +208,13 @@ class MembershipAPITest:
         assert request.handled_by is None
         assert request.refusal_comment is None
 
-    def test_request_existing_pending_membership(self, api):
+    def test_request_existing_pending_membership_do_not_duplicate_it(self, api):
         user = api.login()
         previous_request = MembershipRequest(user=user, comment='previous')
         organization = OrganizationFactory(requests=[previous_request])
         data = {'comment': 'a comment'}
 
-        api_url = url_for('api.request_membership', org=organization)
-        response = api.post(api_url, data)
+        response = api.post(url_for('api.request_membership', org=organization), data)
         assert200(response)
 
         organization.reload()
@@ -232,6 +230,81 @@ class MembershipAPITest:
         assert request.handled_on is None
         assert request.handled_by is None
         assert request.refusal_comment is None
+
+    def test_get_membership_requests(self, api):
+        user = api.login()
+        applicant = UserFactory(email="thibaud@example.org")
+        membership_request = MembershipRequest(user=applicant, comment='test')
+        member = Member(user=user, role='admin')
+        organization = OrganizationFactory(
+            members=[member], requests=[membership_request])
+
+        response = api.get(url_for('api.request_membership', org=organization))
+        assert200(response)
+
+        assert len(response.json) == 1
+        assert response.json[0]['comment'] == 'test'
+        assert response.json[0]['user']['email'] == 'thibaud@example.org' # Can see email of applicant
+
+    def test_only_org_member_can_get_membership_requests(self, api):
+        api.login()
+        applicant = UserFactory(email="thibaud@example.org")
+        membership_request = MembershipRequest(user=applicant, comment='test')
+        organization = OrganizationFactory(
+            members=[], requests=[membership_request])
+
+        response = api.get(url_for('api.request_membership', org=organization))
+        assert403(response)
+
+
+    def test_get_members_with_or_without_email(self, api):
+        admin = Member(user=UserFactory(email="admin@example.org"), role='admin', since="2024-04-14")
+        editor = Member(user=UserFactory(email="editor@example.org"), role='editor')
+        other = UserFactory(email="other@example.org")
+
+        organization = OrganizationFactory(members=[admin, editor])
+
+        # Admin can see emails
+        api.login(admin.user)
+        response = api.get(url_for('api.organization', org=organization))
+        assert200(response)
+
+        members = response.json['members']
+        assert len(members) == 2
+        assert members[0]['role'] == 'admin'
+        assert members[0]['since'] == '2024-04-14T00:00:00+00:00'
+        assert members[0]['user']['email'] == 'admin@example.org'
+
+        assert members[1]['role'] == 'editor'
+        assert members[1]['user']['email'] == 'editor@example.org'
+
+        # Editor can see emails
+        api.login(editor.user)
+        response = api.get(url_for('api.organization', org=organization))
+        assert200(response)
+
+        members = response.json['members']
+        assert len(members) == 2
+        assert members[0]['role'] == 'admin'
+        assert members[0]['since'] == '2024-04-14T00:00:00+00:00'
+        assert members[0]['user']['email'] == 'admin@example.org'
+
+        assert members[1]['role'] == 'editor'
+        assert members[1]['user']['email'] == 'editor@example.org'
+        
+        # Others cannot see emails
+        api.login(other)
+        response = api.get(url_for('api.organization', org=organization))
+        assert200(response)
+
+        members = response.json['members']
+        assert len(members) == 2
+        assert members[0]['role'] == 'admin'
+        assert members[0]['since'] == '2024-04-14T00:00:00+00:00'
+        assert members[0]['user']['email'] is None
+
+        assert members[1]['role'] == 'editor'
+        assert members[1]['user']['email'] is None
 
     def test_accept_membership(self, api):
         user = api.login()